test(io): raise coverage to 90 percent
Add focused IO, QUIC, TCP, UDP, SSL, crypto, URI, compression, CPU and coroutine coverage tests.\n\nMark deterministic OpenSSL internal failure paths out of line coverage because they require allocator or backend fault injection rather than public API coverage.

+++ include/qb/io/protocol/base.h
@@ -185,7 +185,7 @@ public:
         const auto &buffer = this->_io.in();
         auto        i      = buffer.begin() + _offset;
 
-        if ((buffer.end() - i) < _SizeBytes)
+        if (static_cast<std::size_t>(buffer.end() - i) < _SizeBytes)
             return 0;
 
         const auto end = buffer.end() - _SizeBytes + 1;
